@@ -23,61 +23,70 @@
 		"not op_mini all"
 	],
 	"dependencies": {
-		"axios": "^0.21.2",
+		"@babel/eslint-parser": "^7.18.2",
+		"@testing-library/jest-dom": "^5.16.4",
+		"@testing-library/react": "^13.3.0",
+		"@testing-library/user-event": "^14.2.0",
+		"axios": "^0.27.2",
 		"axois": "0.0.1-security",
-		"bootstrap": "^4.6.0",
+		"bootstrap": "^5.1.3",
+		"buffer": "^6.0.3",
 		"deep-object-diff": "^1.1.0",
 		"enzyme": "^3.11.0",
-		"enzyme-adapter-react-16": "^1.15.2",
 		"html-react-parser": "^1.1.2",
-		"jest": "^24.9.0",
+		"jest": "^28.1.0",
 		"jest-enzyme": "^7.1.2",
 		"jest-redux-thunk": "^1.1.1",
 		"jquery": "^3.5.0",
-		"js-cookie": "^2.2.0",
-		"luxon": "^1.23.0",
-		"react": "^16.13.1",
-		"react-bootstrap": "^1.4.3",
-		"react-dnd": "^10.0.2",
-		"react-dnd-html5-backend": "^10.0.2",
-		"react-dnd-test-backend": "^11.1.3",
-		"react-dnd-test-utils": "^11.1.3",
-		"react-dom": "^16.13.1",
-		"react-grid-layout": "^0.18.3",
+		"js-cookie": "^3.0.1",
+		"luxon": "^2.4.0",
+		"process": "^0.11.10",
+		"react": "^18.1.0",
+		"react-bootstrap": "^2.4.0",
+		"react-dnd": "^16.0.1",
+		"react-dnd-html5-backend": "^16.0.1",
+		"react-dnd-test-backend": "^16.0.1",
+		"react-dnd-test-utils": "^16.0.1",
+		"react-dom": "^18.1.0",
+		"react-grid-layout": "^1.3.4",
 		"react-player": "^2.6.0",
 		"react-recaptcha": "^2.3.10",
-		"react-redux": "^7.1.0",
+		"react-redux": "^8.0.2",
 		"react-rnd": "^10.1.10",
-		"react-router": "^5.1.2",
-		"react-router-dom": "^5.1.2",
-		"react-scripts": "^3.4.1",
+		"react-router": "^6.3.0",
+		"react-router-dom": "^6.3.0",
+		"react-scripts": "5.0.1",
 		"react-scrubber": "^1.0.0",
 		"react-uuid": "^1.0.2",
 		"redux": "^4.0.1",
 		"redux-mock-store": "^1.5.4",
 		"redux-thunk": "^2.3.0",
-		"sinon": "^9.2.1",
-		"styled-components": "^4.4.0",
+		"sinon": "^14.0.0",
+		"stream": "0.0.2",
+		"stream-browserify": "^3.0.0",
+		"styled-components": "^5.3.5",
 		"subtitle": "^2.0.3",
-		"sweetalert2": "^10.12.6",
-		"uuid": "^3.3.2",
+		"sweetalert2": "^11.4.17",
+		"uuid": "^8.3.2",
 		"yvideo-editorwidgets": "^1.1.2",
 		"yvideo-subtitle-timeline-editor": "^0.2.7",
 		"yvideojs": "^1.1.8"
 	},
 	"devDependencies": {
-		"@testing-library/dom": "^7.26.0",
-		"eslint-config-react-app": "^5.2.1",
-		"eslint-plugin-flowtype": "^4.7.0",
+		"@cfaester/enzyme-adapter-react-18": "^0.5.1",
+		"@testing-library/dom": "^8.13.0",
+		"eslint-config-react-app": "^7.0.1",
+		"eslint-plugin-flowtype": "^8.0.3",
 		"eslint-plugin-import": "^2.20.2",
 		"eslint-plugin-import-quotes": "^0.0.1",
 		"eslint-plugin-jsx-a11y": "^6.2.3",
 		"eslint-plugin-react": "^7.19.0",
-		"eslint-plugin-react-hooks": "^3.0.0",
-		"husky": "^4.2.5",
+		"eslint-plugin-react-hooks": "^4.5.0",
+		"husky": "^8.0.1",
 		"jest-mock-axios": "^4.2.1",
 		"jest-spy-object": "^1.4.0",
 		"mutation-observer": "^1.0.3",
+		"react-app-rewired": "^2.2.1",
 		"redux-devtools-extension": "^2.13.8"
 	}
 }
