@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import { connect } from 'react-redux'
-import { useHistory } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 import { Tooltip } from 'components/bits'
 
 import {
@@ -32,7 +32,7 @@ const ContentOverviewContainer = props => {
 		toggleTip,
 	} = props
 
-	const history = useHistory()
+	const navigate = useNavigate()
 
 	const [editing, setEditing] = useState(false)
 	const [showing, setShowing] = useState(false)
@@ -191,15 +191,15 @@ const ContentOverviewContainer = props => {
 		const classname = e.target.className
 		if(classname){
 			if(classname.includes(`video-editor`)){
-				history.push({
+				navigate({
 					pathname: `/videoeditor/${content.id}`,
 				})
 			} else if(classname.includes(`subtitle-editor`)){
-				history.push({
-					pathname: `/subtileeditor/${content.id}`,
+				navigate({
+					pathname: `/subtitleeditor/${content.id}`,
 				})
 			} else if(classname.includes(`clip-manager`)){
-				history.push({
+				navigate({
 					pathname: `/clipeditor/${content.id}`,
 				})
 			}
