@@ -3,14 +3,13 @@ import { mount } from 'enzyme'
 import SubtitlesCard from '../../../../components/bits/SubtitlesCard/index'
 import { BrowserRouter} from 'react-router-dom'
 import { DndProvider } from 'react-dnd'
-import Backend from 'react-dnd-html5-backend'
-// import { screen, render, fireEvent } from 'testing-library/react'
+import { HTML5Backend } from 'react-dnd-html5-backend'
 
 describe(`event card`, () => {
 	it(`mock useDrag`, ()=> {
 		const	wrapper = mount(
 			<BrowserRouter>
-				<DndProvider backend={Backend}>
+				<DndProvider backend={HTML5Backend}>
 					<SubtitlesCard />
 				</DndProvider>
 			</BrowserRouter>,
