@@ -1,7 +1,6 @@
 import React from 'react'
 import Tag from '../../../../components/bits/Tag'
-import { Remove } from '../../../../components/bits/Tag/styles'
-import { shallow, mount } from 'enzyme'
+import { render, cleanup, screen, fireEvent } from '@testing-library/react'
 import { BrowserRouter} from 'react-router-dom'
 
 const props = {
@@ -27,25 +26,35 @@ const props = {
 	clientUser: { id: `user:2` },
 	client: { id: `test`, name: `test` },
 	content: { files: [[Object]] },
+	onClick: jest.fn(),
+	children: `test`,
 
 }
 
-const handlers = {
+const handlers = { // eslint-disable-line no-unused-vars
 	removeTag: jest.fn(),
 	removeWord: jest.fn(),
 }
 
+const wrapper =
+<BrowserRouter>
+	<Tag {...props} />
+</BrowserRouter>
+
 describe(`tag test`, () => {
-	it(`mou removeTag`, () => {
-		const button = shallow(<Remove onClick={handlers.removeTag} data-value={props} />)
-		button.find(`StyledComponent`).simulate(`click`)
-		expect(handlers.removeTag.mock.calls.length).toEqual(1)
+	beforeEach(() => {
+		render(wrapper)
+	})
+	afterEach(() => {
+		jest.resetAllMocks()
+		cleanup()
 	})
 	it(`mou removeTag`, () => {
-		mount(
-			<BrowserRouter>
-				<Tag {...props} />
-			</BrowserRouter>,
-		)
+		const Remove = screen.getByTestId(`remove`)
+		fireEvent.click(Remove)
+
+		expect(props.onClick).toHaveBeenCalledTimes(1)
+		expect(screen.getByText(/test/)).toBeVisible()
+
 	})
 })
