@@ -20,14 +20,26 @@ const viewstate = {
 		isLabAssistantRoute:false,
 	},
 }
-
+// routeProps.children.props.children.props.
 describe(`root route paring test`, () => {
 	it(`should be true`, ()=> {
 		const root = shallow(<Root viewstate={viewstate}/>)
 		const pathMap = root.find(Route).reduce((pathMap, route) => {
 			const routeProps = route.props()
-			pathMap[routeProps.path] = routeProps.children.type.WrappedComponent.name
+			if (routeProps.children === undefined) {
+				pathMap[routeProps.path] =
+					routeProps.element.type.WrappedComponent.name
+			} else {
+				if (routeProps.children.props.children === undefined) {
+					pathMap[`${routeProps.path}/${routeProps.children.props.path}`] =
+						routeProps.element.type.WrappedComponent.name
+				} else {
+					pathMap[`${routeProps.path}/${routeProps.children.props.path}/${routeProps.children.props.children.props.path}`] =
+							routeProps.element.type.WrappedComponent.name
+				}
+			}
 			return pathMap
+
 		}, {})
 
 		expect(pathMap[`/`]).toBe(`CollectionsContainer`)
@@ -35,12 +47,12 @@ describe(`root route paring test`, () => {
 		expect(pathMap[`/collections`]).toBe(`CollectionsContainer`)
 		expect(pathMap[`/lab-assistant`]).toBe(`LabAssistantContainer`)
 		expect(pathMap[`/manage-resource`]).toBe(`ManageResourceContainer`)
-		expect(pathMap[`/lab-assistant-manager/:professorId/:collectionId?`]).toBe(`LabAssistantManagerContainer`)
-		expect(pathMap[`/manager/:id?`]).toBe(`ManagerContainer`)
-		expect(pathMap[`/public-manager/:id?`]).toBe(`PublicManagerContainer`)
-		expect(pathMap[`/player/:id/:clip?`]).toBe(`PlayerContainer`)
+		expect(pathMap[`/lab-assistant-manager/:professorId/:collectionId`]).toBe(`LabAssistantManagerContainer`)
+		expect(pathMap[`/manager/:id`]).toBe(`ManagerContainer`)
+		expect(pathMap[`/public-manager/:id`]).toBe(`PublicManagerContainer`)
+		expect(pathMap[`/player/:id/:clip`]).toBe(`PlayerContainer`)
 		expect(pathMap[`/videoeditor/:id`]).toBe(`VideoEditorContainer`)
-		expect(pathMap[`/subtileeditor/:id`]).toBe(`SubtitlesEditorContainer`)
+		expect(pathMap[`/subtitleeditor/:id`]).toBe(`SubtitlesEditorContainer`)
 		expect(pathMap[`/clipeditor/:id`]).toBe(`ClipEditorContainer`)
 		expect(pathMap[`/feedback`]).toBe(`FeedbackContainer`)
 		expect(pathMap[`/search-public-collections`]).toBe(`SearchPublicCollectionsContainer`)
