@@ -17,7 +17,7 @@ const SubtitlesCard = (props) => {
 			{
 				isEdit ?
 					subLayer === index ?
-						<input type='text' className='sideTabInput' style={{margin: `0px`, width: `100%`}} value={value}
+						<input data-testid='title-change' type='text' className='sideTabInput' style={{margin: `0px`, width: `100%`}} value={value}
 							onChange={handleChange}
 							onKeyPress={event => {
 								if (event.charCode === 13)
