@@ -48,10 +48,14 @@ const SubtitlesLayer = props => {
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [width])
 
-	if(document.getElementsByClassName(`total`)[0] !== undefined && layerWidth !== 0){
-		document.getElementById(`time-bar-container`).style.width = `${layerWidth - 163}px`
-		document.getElementsByClassName(`total`)[0].style.width = `${layerWidth - 163}px`
-		document.getElementById(`layer-time-indicator`).style.width = `${layerWidth}px`
+	const tempOnload = window.onload
+	window.onload = () => {
+		if(document.getElementsByClassName(`total`)[0] !== undefined && layerWidth !== 0) {
+			document.getElementById(`time-bar-container`).style.width = `${layerWidth - 163}px`
+			document.getElementsByClassName(`total`)[0].style.width = `${layerWidth - 163}px`
+			document.getElementById(`layer-time-indicator`).style.width = `${layerWidth}px`
+		}
+		window.onload = tempOnload
 	}
 	// This object is to tell the onReziseStop nevent for the Rnd component that resizing can only be right and left
 	const Enable = {
