@@ -44,13 +44,15 @@ const TrackLayer = props => {
 			// we are in censor, calculate overlapping
 			// overlap count tells us how many half layers we need
 			const overlapCount = calculateOverlaps()
-
-			if(overlapCount.length !== layerOverlap.length) setLayerOverlap(overlapCount)
-
-			document.getElementById(`layer-${layerIndex}`).style.height =
-			`${overlapCount.length !== 0 ?
-				26 * (overlapCount.length + 1)
-				: 46}px`
+			const tempOnload = window.onload
+			window.onload = () => {
+				if(overlapCount.length !== layerOverlap.length) setLayerOverlap(overlapCount)
+				document.getElementById(`layer-${layerIndex}`).style.height =
+				`${overlapCount.length !== 0 ?
+					26 * (overlapCount.length + 1)
+					: 46}px`
+				window.onload = tempOnload
+			}
 		}
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [width, events, layerOverlap])
@@ -65,9 +67,9 @@ const TrackLayer = props => {
 			setLayerWidth(layerWidth + width)
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	},[width])
-	if(document.getElementsByClassName(`total`)[0] !== undefined && layerWidth !== 0){
+	if(document.getElementById(`timeBarProgress`) !== undefined && layerWidth !== 0){
 		document.getElementById(`time-bar-container`).style.width = `${layerWidth - 2}px`
-		document.getElementsByClassName(`total`)[0].style.width = `${layerWidth - 2}px`
+		document.getElementById(`timeBarProgress`).style.width = `${layerWidth - 2}px`
 		document.getElementById(`layer-time-indicator`).style.width = `${layerWidth}px`
 	}
 	// This object is to tell the onReziseStop nevent for the Rnd component that resizing can only be right and left
@@ -254,10 +256,10 @@ const TrackLayer = props => {
 							className={`layer-${layerIndex} ${layerOverlap.length > 0 ? `half-layer` : ``} events ${displayLayer === layerIndex ? `active-layer` : ``}`}
 							style={
 								{
+									backgroundColor: `rgba(5, 130, 202, 0.1)`,
 									marginTop: layerOverlap.length > 0 ?
 										`${26 * layerOverlap.length}px`
 										: `0px`,
-									backgroundColor: `rgba(5, 130, 202, 0.1)`,
 								}
 							}
 						>
@@ -283,7 +285,8 @@ const TrackLayer = props => {
 											: `0px`,
 										backgroundColor: overlapIndex % 2 !== 0 ?
 											`rgba(5, 130, 202, 0.1)`
-											: ``,
+											:
+											``,
 									}}
 							>
 								{
