@@ -103,7 +103,11 @@ const VideoContainer = props => {
 			}
 
 			setElapsed(playedSeconds)
-			document.getElementById(`seconds-time-holder`).innerText = playedSeconds
+			const tempOnload = window.onload
+			window.onload = () => {
+				document.getElementById(`seconds-time-holder`).innerText = playedSeconds
+				window.onload = tempOnload
+			}
 
 			if(!events) return
 			const values = CurrentEvents(playedSeconds,events,duration)
@@ -332,6 +336,7 @@ const VideoContainer = props => {
 	}
 
 	let count = 0 // this is to make sure that event listeners are applied only once
+	let isPlaying = false
 
 	const handleHotKeys = (e) => { // eslint-disable-line no-unused-vars
 		const playedTime = parseFloat(document.getElementById(`seconds-time-holder`).innerHTML)
@@ -349,13 +354,13 @@ const VideoContainer = props => {
 			video.handleSeek(null, playedTime - .1)
 			break
 		case `Space`:
-			setPlaying(playing)
-			if (playing === true)
+			if(isPlaying) {
 				video.handlePause()
-
-			if (playing === false)
+				isPlaying = false
+			} else {
 				video.handlePlay()
-
+				isPlaying = true
+			}
 			break
 
 		default:
@@ -368,29 +373,33 @@ const VideoContainer = props => {
 			count++
 			// checking for time bar and setting event listener
 			if(document.getElementById(`time-bar`) !== null && duration !== 0){
-				document.getElementById(`time-bar`).addEventListener(`mousemove`, (e) => {
-					// calculate current time based on mouse position
-					const currentLayerWidth = document.getElementById(`time-bar-container`).clientWidth
-					const currentScrollLeft = document.getElementById(`time-bar-container`).scrollLeft
+				const tempOnload = window.onload
+				window.onload = () => {
+					document.getElementById(`time-bar`).addEventListener(`mousemove`, (e) => {
+						// calculate current time based on mouse position
+						const currentLayerWidth = document.getElementById(`time-bar-container`).clientWidth
+						const currentScrollLeft = document.getElementById(`time-bar-container`).scrollLeft
 
-					const secondsCurrentTimePercent = (e.offsetX + currentScrollLeft) / currentLayerWidth
+						const secondsCurrentTimePercent = (e.offsetX + currentScrollLeft) / currentLayerWidth
 
-					const dateElapsed = new Date(null)
-					dateElapsed.setSeconds(secondsCurrentTimePercent * duration)
-					const formattedElapsed = dateElapsed.toISOString().substr(11, 8)
+						const dateElapsed = new Date(null)
+						dateElapsed.setSeconds(secondsCurrentTimePercent * duration)
+						const formattedElapsed = dateElapsed.toISOString().substr(11, 8)
 
-					// set new x position to the red bar
-					document.getElementById(`time-bar-shadow`).style.visibility = `visible`
-					document.getElementById(`time-bar-shadow`).style.transform = `translateX(${e.offsetX - 2}px)`
-					document.getElementById(`time-bar-shadow-text`).innerText = `${formattedElapsed}`
-					if(e.offsetX > window.innerWidth / 2)
-						document.getElementById(`time-bar-shadow-text`).style.right = `6rem`
-					else
-						document.getElementById(`time-bar-shadow-text`).style.right = `0`
+						// set new x position to the red bar
+						document.getElementById(`time-bar-shadow`).style.visibility = `visible`
+						document.getElementById(`time-bar-shadow`).style.transform = `translateX(${e.offsetX - 2}px)`
+						document.getElementById(`time-bar-shadow-text`).innerText = `${formattedElapsed}`
+						if(e.offsetX > window.innerWidth / 2)
+							document.getElementById(`time-bar-shadow-text`).style.right = `6rem`
+						else
+							document.getElementById(`time-bar-shadow-text`).style.right = `0`
 
-					document.getElementById(`layer-time-indicator-line-shadow`).style.visibility = `visible`
-					document.getElementById(`layer-time-indicator-line-shadow`).style.transform = `translateX(${e.offsetX}px)`
-				})
+						document.getElementById(`layer-time-indicator-line-shadow`).style.visibility = `visible`
+						document.getElementById(`layer-time-indicator-line-shadow`).style.transform = `translateX(${e.offsetX}px)`
+						window.onload = tempOnload
+					})
+				}
 			}
 			// checking video container and setting event listener for hot keys
 			window.onkeyup = (e) => {
@@ -465,7 +474,7 @@ const VideoContainer = props => {
 			{/* Load the spinner if a file is uploading */}
 			{!isUploading && <div className='loading-spinner'><Spinner/></div> }
 
-			<ReactPlayer ref={ref} config={config} url={url}
+			<ReactPlayer ref={ref} config={config} url={url} data-testid='react-player'
 				onContextMenu={e => e.preventDefault()}
 				key={url}
 
@@ -510,8 +519,12 @@ const VideoContainer = props => {
 
 						<div id='time-bar' onMouseLeave={(e) => {
 							if(document.getElementById(`time-bar-shadow`) !== null && document.getElementById(`layer-time-indicator-line-shadow`) !== null) {
-								document.getElementById(`time-bar-shadow`).style.visibility = `hidden`
-								document.getElementById(`layer-time-indicator-line-shadow`).style.visibility = `hidden`
+								const tempOnload = window.onload
+								window.onload = () => {
+									document.getElementById(`time-bar-shadow`).style.visibility = `hidden`
+									document.getElementById(`layer-time-indicator-line-shadow`).style.visibility = `hidden`
+									window.onload = tempOnload
+								}
 							}
 						}}>
 							<div id={`time-bar-container`}>
