@@ -17,7 +17,7 @@ const SubtitlesContainer = props => {
 		}else
 			setSubtitlesArray([])
 
-		for(let i = 0; i<subtitlesArray.length; i++){
+		for(let i = 0; i < subtitlesArray.length; i++){
 			const element = subtitlesArray[i]
 			const start = element.start
 			const end = element.end
@@ -27,7 +27,6 @@ const SubtitlesContainer = props => {
 			}else if (currentTime > end || currentTime < start)
 				handleShowSubtitle(``)
 		}
-
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [duration, subtitles])
 
