@@ -1,7 +1,13 @@
 import Enzyme from 'enzyme'
-import EnzymeAdapter from 'enzyme-adapter-react-16'
+import Adapter from '@cfaester/enzyme-adapter-react-18'
+import { configure } from '@testing-library/dom'
+import '@testing-library/jest-dom'
 
 Enzyme.configure({
-	adapter: new EnzymeAdapter(),
+	adapter: new Adapter(),
 	disableLifecycleMethods: true,
+})
+
+configure({
+	asyncUtilTimeout: 2500,
 })
